Use Checkstyle
Checkstyle is a development tool to blame about coding style failures.
Therefore several plugins are required in build and report section of the
project. With the usage of the pluginManagement section in build section,
it is ensured, that same version of plugins is used for build and
reporting.

checkstyle project is licensed under LGPL-2.1. Therefore LGPL-2.1 SPDX
license file is added.

The config/checkstyle/checkstyle.xml file is based on the
src/main/resources/sun_checks.xml of the official checkstyle project.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>
Reviewed-by: Thomas Gleixner <tglx@linutronix.de>